Move fields() into nebula, return as nebula data
