+++ hubblestack/extmods/modules/nebula_osquery.py
@@ -198,6 +198,29 @@ def queries(query_group,
     return ret
 
 
+def fields(custom_fields=None):
+    '''
+    Use config.get to retrieve custom data based on the keys in the `custom_fields`
+    list.
+
+    Arguments:
+
+    custom_fields
+        List of keys to retrieve
+    '''
+    if custom_fields is None or not isinstance(fields, list):
+        log.error('custom_fields argument must be formed as a list of strings')
+        return []
+    ret = {}
+    for field in custom_fields:
+        ret[field] = __salt__['config.get'](field)
+    # Return it as nebula data
+    return [{'custom_fields': {
+                 'data': [ret],
+                 'result': True
+            }}]
+
+
 def version():
     '''
     Report version of this module
